refactor(http): move BaseController to base-controller.ts
