+++ packages/http/src/index.ts
@@ -3,6 +3,9 @@ export * from './response-methods/index.js';
 export * from './decorators.js';
 export * from './responses.js';
 export * from './route-args/index.js';
+export * from './exceptions.js';
+export * from './base-controller.js';
+export * from './route-builder.js';
 export * from './controllers.js';
 export * from './server.js';
 export * from './transformers/index.js';
